+++ packaging/guix.scm
@@ -38,7 +38,7 @@
 
 (package
   (name "torando-gui")
-  (version "1.3.2")
+  (version "1.3.3")
   (source (local-file %source "torando-gui-checkout"
                       #:recursive? #t
                       #:select? source-file?))